fix: code health improvements from PR review audit (SIMPLE-28)
Address all 8 subtasks from the Phase 3 code health audit:

- SIMPLE-35: Split 4 files below 800-line guideline
  - s3impl.rs 988→798 (extract helpers to s3impl_helpers.rs)
  - grpc.rs 991→798 (move logic to grpc_helpers.rs)
  - storage/mod.rs 882→440 (extract segment I/O to segment.rs)
  - serve.rs 838→778 (extract lifecycle to serve_lifecycle.rs)
- SIMPLE-36: Extract run() phases into spawn_background_tasks,
  validate_metrics_auth, graceful_shutdown
- SIMPLE-37: Extract delete_one_versioned, delete_one_unversioned,
  build_version_lists helpers from long functions
- SIMPLE-38: Add test_head_bucket_auth_failure test
- SIMPLE-39: Pin minio/mint:edge to SHA digest, ceph/s3-tests to commit
- SIMPLE-40: Fix tenant credential reuse (use alt_key/alt_secret)
- SIMPLE-41: Pre-build tox virtualenv in ceph s3-tests Dockerfile
- SIMPLE-42: Clarify CRC32C conditionality in backup docs

Also fixes bench compilation errors (append_data 4-tuple, put_meta &mut).
