+++ docker/ceph-s3-tests/Dockerfile
@@ -1,6 +1,6 @@
 FROM python:3.12-slim
 
-ARG S3_TESTS_REF=master
+ARG S3_TESTS_REF=06e2c5789a39991eed6008b8f728914fe7c7ee3a
 ARG TOX_VERSION=4.25.0
 
 RUN apt-get update && apt-get install -y --no-install-recommends \
@@ -12,6 +12,7 @@ RUN git clone https://github.com/ceph/s3-tests.git /s3-tests && \
 
 WORKDIR /s3-tests
 RUN pip install --no-cache-dir "tox==${TOX_VERSION}"
+RUN tox --notest -e py
 
 RUN useradd --create-home --uid 10001 s3tests && \
     chown -R s3tests:s3tests /s3-tests
